requirements.txt
python-telegram-bot==13.14
httpx==0.23.3
mypy==1.1.1
isort==5.12.0
wemake-python-styleguide==0.17.0
pydantic==1.10.7
pytest==7.2.2
pytest-cov==4.0.0
jinja2==3.1.2
